Yugioh card component naming changes and simplified usage
