+++ src/components/card/YGOCardStyles.js
